Convert GRAS pid from long int to int; setup a virtualization layer in XBT to choose between GRAS and MSG when searching to compute the pid, for example in the logs
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@3643 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ include/gras/virtu.h
@@ -49,7 +49,7 @@ XBT_PUBLIC(int) gras_os_myport(void);
  * Returns the process ID of the current process.  (This is often used
    by routines that generate unique temporary file names.)
  */
-XBT_PUBLIC(long int) gras_os_getpid(void);
+XBT_PUBLIC(int) gras_os_getpid(void);
 
 /** @} */
 SG_END_DECL()
